refactor: Single tree for `RootMeasurementWriter`
